First commit. More information below.
Sokkit is an API that interfaces with both Bukkit and Bungee (coming to
Trident and Sponge when released) that allows for easy inter-server
communication via Sockets.

It allows for Authentication, and has a defined protocol that ensures
nothing is misinterpreted.

Sokkit is multi-threaded to ensure it does not hang the main thread.